pyproject.toml
[tool.poetry]
name = "shortlink"
version = "0.1.0"
description = ""
authors = ["nikita-r <gfgfgf354@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
